+++ bob.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  ignore: ['@graphql-mesh/container', '@graphql-mesh/playground'],
+  ignore: ['@graphql-mesh/container', '@graphql-mesh/cross-helpers'],
 };