+++ packages/cross-helpers/browser.js
@@ -1,0 +1,4 @@
+module.exports = {
+  fs: require('browserify-fs'),
+  path: require('path-browserify'),
+};