+++ packages/cross-helpers/index.d.ts
@@ -1,0 +1,4 @@
+import fs from 'fs';
+import path from 'path';
+
+export { fs, path };