+++ examples/graphql-file-upload-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # graphql-file-upload-example
 
+## 0.6.168
+
+### Patch Changes
+
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/cli@0.64.0
+  - @graphql-mesh/graphql@0.22.4
+
+## 0.6.167
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/cli@0.63.0
+  - @graphql-mesh/graphql@0.22.3
+
 ## 0.6.166
 
 ### Patch Changes