+++ examples/json-schema-covid/CHANGELOG.md
@@ -1,5 +1,29 @@
 # covid-mesh
 
+## 0.5.224
+
+### Patch Changes
+
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/cli@0.64.0
+  - @graphql-mesh/graphql@0.22.4
+  - @graphql-mesh/json-schema@0.26.5
+
+## 0.5.223
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/cli@0.63.0
+  - @graphql-mesh/graphql@0.22.3
+  - @graphql-mesh/json-schema@0.26.4
+
+## 0.5.222
+
+### Patch Changes
+
+- @graphql-mesh/json-schema@0.26.3
+
 ## 0.5.221
 
 ### Patch Changes