+++ examples/odata-microsoft/CHANGELOG.md
@@ -1,5 +1,21 @@
 # odata-microsoft-graph-example
 
+## 0.6.181
+
+### Patch Changes
+
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/cli@0.64.0
+  - @graphql-mesh/odata@0.17.1
+
+## 0.6.180
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/cli@0.63.0
+  - @graphql-mesh/odata@0.17.0
+
 ## 0.6.179
 
 ### Patch Changes