+++ examples/openapi-youtrack/CHANGELOG.md
@@ -1,5 +1,23 @@
 # openapi-youtrack
 
+## 0.5.207
+
+### Patch Changes
+
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/cli@0.64.0
+  - @graphql-mesh/openapi@0.24.1
+  - @graphql-mesh/transform-resolvers-composition@0.12.34
+
+## 0.5.206
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/cli@0.63.0
+  - @graphql-mesh/openapi@0.24.0
+  - @graphql-mesh/transform-resolvers-composition@0.12.33
+
 ## 0.5.205
 
 ### Patch Changes