+++ examples/sqlite-chinook/CHANGELOG.md
@@ -1,5 +1,21 @@
 # chinook
 
+## 0.5.194
+
+### Patch Changes
+
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/cli@0.64.0
+  - @graphql-mesh/tuql@0.5.1
+
+## 0.5.193
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/cli@0.63.0
+  - @graphql-mesh/tuql@0.5.0
+
 ## 0.5.192
 
 ### Patch Changes