+++ packages/cache/redis/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @graphql-mesh/cache-redis
 
+## 0.7.25
+
+### Patch Changes
+
+- Updated dependencies [b1a6df928]
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/types@0.68.2
+  - @graphql-mesh/utils@0.32.0
+  - @graphql-mesh/cache-inmemory-lru@0.6.4
+
+## 0.7.24
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/utils@0.31.0
+  - @graphql-mesh/types@0.68.1
+  - @graphql-mesh/cache-inmemory-lru@0.6.3
+
 ## 0.7.23
 
 ### Patch Changes