+++ packages/cache/store/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @graphql-mesh/cache-store
 
+## 0.1.58
+
+### Patch Changes
+
+- Updated dependencies [b1a6df928]
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/types@0.68.2
+  - @graphql-mesh/store@0.6.0
+
+## 0.1.57
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/store@0.5.0
+  - @graphql-mesh/types@0.68.1
+
 ## 0.1.56
 
 ### Patch Changes