+++ packages/handlers/json-schema/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @graphql-mesh/json-schema
 
+## 0.26.5
+
+### Patch Changes
+
+- Updated dependencies [b1a6df928]
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/types@0.68.2
+  - json-machete@0.7.0
+  - @graphql-mesh/store@0.6.0
+  - @graphql-mesh/utils@0.32.0
+  - @omnigraph/json-schema@0.17.13
+
+## 0.26.4
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/store@0.5.0
+  - @graphql-mesh/utils@0.31.0
+  - @graphql-mesh/types@0.68.1
+  - json-machete@0.6.1
+  - @omnigraph/json-schema@0.17.12
+
+## 0.26.3
+
+### Patch Changes
+
+- Updated dependencies [b45dac0c0]
+- Updated dependencies [b45dac0c0]
+  - json-machete@0.6.0
+  - @omnigraph/json-schema@0.17.11
+
 ## 0.26.2
 
 ### Patch Changes