+++ packages/handlers/tuql/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @graphql-mesh/tuql
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [b1a6df928]
+  - @graphql-mesh/types@0.68.2
+
+## 0.5.0
+
+### Minor Changes
+
+- b2c537c2a: feat - cross-platform support
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/cross-helpers@0.1.0
+  - @graphql-mesh/types@0.68.1
+
 ## 0.4.104
 
 ### Patch Changes