+++ packages/json-machete/CHANGELOG.md
@@ -1,5 +1,29 @@
 # json-machete
 
+## 0.7.0
+
+### Minor Changes
+
+- 67fb11706: enhance: improve cross-platform support
+
+### Patch Changes
+
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/utils@0.32.0
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/utils@0.31.0
+
+## 0.6.0
+
+### Minor Changes
+
+- b45dac0c0: feat(json-machete): add missing format if example is provided
+
 ## 0.5.20
 
 ### Patch Changes