+++ packages/mergers/stitching/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @graphql-mesh/merger-stitching
 
+## 0.15.28
+
+### Patch Changes
+
+- Updated dependencies [b1a6df928]
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/types@0.68.2
+  - @graphql-mesh/store@0.6.0
+  - @graphql-mesh/utils@0.32.0
+
+## 0.15.27
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/store@0.5.0
+  - @graphql-mesh/utils@0.31.0
+  - @graphql-mesh/types@0.68.1
+
 ## 0.15.26
 
 ### Patch Changes