+++ packages/transforms/rate-limit/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @graphql-mesh/transform-rate-limit
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [b1a6df928]
+- Updated dependencies [67fb11706]
+  - @graphql-mesh/types@0.68.2
+  - @graphql-mesh/utils@0.32.0
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [b2c537c2a]
+  - @graphql-mesh/utils@0.31.0
+  - @graphql-mesh/types@0.68.1
+
+## 0.1.1
+
+### Patch Changes
+
+- e23ad8237: fix(rate-limit): do not break the request in case of nested fields
+
 ## 0.1.0
 
 ### Minor Changes